WIP: enable typescript@next path mapping
This is a hack to make it possible to use the `baseUrl` property
in typescript@next. The `paths` object property may also work, too.
Less confident about the `rootDirs` array property.

I believe that baseUrl is to be interpreted as from the directory
where tsconfig.json lives, and this isn't generally avaialable in
gulp-typescript's API, so one reason (among many) that this is a
hack and not a solution is that it assumes the directory found to
be the 'commonBasePath' is assumed to be the tsconfig file's
directory, which may not always be the case.