Ensure uniqueness of reported errors
The uniqueness of the errors is checked by hashing them and adding
them to a set. If an error is already in the set, it is not reported.
To support older versions of Node, the set is implemented using
the keys of a JS object. The hash of diagnostic is created using a
JSON serializing of object (after dropping circular references).

Closes ivogabe/gulp-typescript#543