Pass the full path of the configuration file.
Fixes #563

The fix is actually the change in main.ts. The same change was applied to
project.ts though I've not found a way to create a test case that generates
there. (It may not be currently possible to generate such a case.) It just seems
that the two calls to parseJsonConfigFileContent should pass the whole path to
the config.